Split up the backend so it's not in a single file
