+++ .gitignore
@@ -2,3 +2,4 @@ server/venv
 **/.DS_Store
 server/.env
 .vscode/*
+**/__pycache__
