+++ client/mcbb-frontend/src/routes/dashboard/addeditinterestpage/addeditinterest.route.tsx
@@ -8,7 +8,9 @@ const AddedInterest = () => {
   const [error, setError] = useState<string[]>([]);
   const [successMessage, setSuccessMessage] = useState('');
 
-  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>): Promise<void> => {
+  const handleSubmit = async (
+    event: React.FormEvent<HTMLFormElement>
+  ): Promise<void> => {
     event.preventDefault();
     setError([]);
     setSuccessMessage('');
@@ -19,13 +21,16 @@ const AddedInterest = () => {
     }
 
     try {
-      const response = await fetch('http://127.0.0.1:3000/api/add_tag', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ tag_name: interestName }),
-      });
+      const response = await fetch(
+        'http://127.0.0.1:3000/api/interests/add-tag',
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify({ tag_name: interestName })
+        }
+      );
 
       const result = await response.json();
 
@@ -51,13 +56,16 @@ const AddedInterest = () => {
     }
 
     try {
-      const response = await fetch('http://127.0.0.1:3000/api/remove_tag', {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ tag_name: interestName }),
-      });
+      const response = await fetch(
+        'http://127.0.0.1:3000/api/interests/remove-tag',
+        {
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify({ tag_name: interestName })
+        }
+      );
 
       const result = await response.json();
 
@@ -75,35 +83,38 @@ const AddedInterest = () => {
 
   return (
     <ResponsiveForm onSubmit={handleSubmit}>
-      <h1 className="text-2xl font-bold text-center">Add Interest</h1>
+      <h1 className='text-2xl font-bold text-center'>Add Interest</h1>
       {error.length > 0 && (
-        <div className="text-red-500">
+        <div className='text-red-500'>
           {error.map((err, idx) => (
             <div key={idx}>{err}</div>
           ))}
         </div>
       )}
-      {successMessage && <div className="text-green-500">{successMessage}</div>}
+      {successMessage && <div className='text-green-500'>{successMessage}</div>}
       <Input
-        label="Interest Name:"
-        name="interestName"
-        type="text"
+        label='Interest Name:'
+        name='interestName'
+        type='text'
         value={interestName}
         onChange={(e) => setInterestName((e.target as HTMLInputElement).value)}
-        placeholder="Enter the interest name"
-        color="blue"
+        placeholder='Enter the interest name'
+        color='blue'
         filled={false}
         required
       />
-      <div className="flex flex-row gap-2 mt-4">
-        <Button text="Add" color="blue" type="submit" filled name="add" />
-        <Button text="Remove" color="red" filled={false} onClick={handleRemove} name="remove" />
+      <div className='flex flex-row gap-2 mt-4'>
+        <Button text='Add' color='blue' type='submit' filled name='add' />
+        <Button
+          text='Remove'
+          color='red'
+          filled={false}
+          onClick={handleRemove}
+          name='remove'
+        />
       </div>
     </ResponsiveForm>
   );
 };
 
 export default AddedInterest;
-
-
-
