+++ client/mcbb-frontend/src/routes/dashboard/club/delete/DeleteClub.loader.tsx
@@ -18,7 +18,7 @@ const deleteClubLoader: LoaderFunction = async ({ params }) => {
     return redirect('/dashboard/clubs');
   }
   const response = await fetch(
-    `http://localhost:3000/api/delete-club/${params.id}`,
+    `http://localhost:3000/api/clubs/delete-club/${params.id}`,
     {
       method: 'DELETE',
       credentials: 'include',
