+++ client/mcbb-frontend/src/routes/dashboard/event/Event.action.tsx
@@ -7,7 +7,7 @@ const eventAction: ActionFunction = async ({ request }) => {
   if (action === 'rsvp') {
     const type = formData.get('type');
     const response = await fetch(
-      `http://localhost:3000/api/rsvp?event_id=${id}&type=${type}`,
+      `http://localhost:3000/api/rsvp/rsvp?event_id=${id}&type=${type}`,
       {
         method: 'POST',
         headers: {
