+++ client/mcbb-frontend/src/routes/dashboard/home/Home.loader.tsx
@@ -11,7 +11,7 @@ const homeLoader: LoaderFunction = async () => {
     return redirect('/verifyEmail');
   }
   const response = await fetch(
-    `http://localhost:3000/api/events?start_date=${encodeURIComponent(
+    `http://localhost:3000/api/events/events?start_date=${encodeURIComponent(
       new Date().toISOString()
     )}&end_date=${encodeURIComponent(
       new Date(
