+++ client/mcbb-frontend/src/routes/logout/Logout.loader.tsx
@@ -1,7 +1,7 @@
 import { redirect } from 'react-router-dom';
 
 const logoutLoader = async () => {
-  await fetch('http://localhost:3000/api/logout', {
+  await fetch('http://localhost:3000/api/auth/logout', {
     method: 'POST',
     credentials: 'include',
     headers: {
