+++ server/extensions.py
@@ -1,0 +1,5 @@
+from flask_mysqldb import MySQL
+from flask_cors import CORS
+
+mysql = MySQL()
+cors = CORS()
